+++ core/language/pt_br/gallery_acp.php
@@ -334,7 +334,6 @@ $lang = array_merge($lang, [
 	'RRC_GINDEX'                 => 'Recurso de imagens recentes, aleatórias e comentários',
 	'RRC_GINDEX_COLUMNS'         => 'Colunas',
 	'RRC_GINDEX_COMMENTS'        => 'Recolher comentários',
-	'RRC_GINDEX_CONTESTS'        => 'Número de concursos',
 	'RRC_GINDEX_CROWS'           => 'Número de comentários',
 	'RRC_GINDEX_MODE'            => 'Modo',
 	'RRC_GINDEX_MODE_EXP'        => 'Em bancos de dados grandes, as imagens aleatórias podem demorar para carregar.',