+++ core/migrations/release_1_2_0.php
@@ -226,7 +226,6 @@ class release_1_2_0 extends migration
 		'prune_orphan_time'		=> 0,
 
 		'rrc_gindex_comments'	=> false,
-		//'rrc_gindex_contests'	=> 1,
 		'rrc_gindex_display'	=> 173,
 		'rrc_gindex_mode'		=> 7,
 		'rrc_gindex_pegas'		=> true,